Split validation tests from scoring tests
