+++ packages/perseus/src/widgets/categorizer/score-categorizer.test.ts
@@ -30,19 +30,4 @@ describe("scoreCategorizer", () => {
 
         expect(score).toHaveBeenAnsweredIncorrectly();
     });
-
-    it("tells the learner its not complete if not selected", () => {
-        const rubric: PerseusCategorizerRubric = {
-            values: [1, 3],
-        };
-
-        const userInput = {
-            values: [2],
-        } as const;
-        const score = scoreCategorizer(userInput, rubric, mockStrings);
-
-        expect(score).toHaveInvalidInput(
-            "Make sure you select something for every row.",
-        );
-    });
 });
